metrics: prob: fix typo in REL and RES
Fix a typo in the code for REL and RES: rather than averaging over all
the forecasts, the code was (incorrectly) averaging by dividing by the
number of unique forecasts.

Also, revise the rounding rule to what was discussed in a GH issue,
i.e., if there are less than 1000 samples, then round to tens (e.g.
0.1, 0.2, etc.); if greater than 1000 samples, then round to hundres
(e.g. 0.01, 0.02, 0.03, etc.).