metrics: switch back to `ref` for reference forecasts
Switch back to using `ref` for reference forecasts (instead of
`fx_ref`). This change is partly due to the fact that the probabilistic
forecasts use `ref` and `ref_prob` (i.e. if we changed to `fx_ref` in
deterministic, the probabilistic metrics would need to change to
`fx_ref` and `fx_ref_prob`).