metrics: preprocess: lower resample threshold
Lower the count threshold in the resampling (to decide whether there is
enough data per interval to consider the resample as valid) to 10% in
order to force the tests to pass.

The issue is that, with the more robust sampling (using the count
thresholding), the `closed="right"` tests fail as there is only one
sample in the "first" hour (11:01 to 12:00) since the test data starts
at 12:00. Due to the way the tests are written, reducing the threshold
solves the failing test issue with the minimal amount of code change.
The alternative (which may be added in a follow-up commit) is to revise
the tests, which would be more robust.