conftest: report_objects: switch from 1h to 15min interval_length
Due to the changes in the `metrics.preprocessing` (making the resampling
more robust), we needed to change the interval_length from 1min to 1h to
make the tests pass (since otherwise too many values would be dropped).
But while 1h worked, it's too conservative of a value. Instead, this
commit switches to 15min, which is long enough to have the tests pass,
but still short enough for other testing.