metrics: add summary statistics (#294)
* metrics: add summary statistics

Add functions to compute summary statistics of the observations and
forecasts, where the functions are either wrappers around numpy
functions or consistent with numpy logic.

* metrics: summary: remove old comments

* metrics: summary: simplify code

Instead of adding wrappers to numpy functions, switch to defining the
available summary functions via the `_MAP` variable (same concept/setup
as in the other metric functions).

* metrics: summary: correct longhand name

* metrics: summary: add simple tests

Add some (admittedly) simple tests to just confirm the code logic works
as expected. In this case, we're not wrapping the numpy functions, but
instead calling them directly via the `_MAP` variable (which matches
what's been done in the deterministic, probabilistic and valuation
metrics).