Switch back to explicit fx/obs generation
To help with debugging, switch back to explicitly defining the fx and
obs objects. Also, it seems there's some sort of bug where the global
variable `DETERMINISTIC_METRICS` doesn't include forecast skill (s) in
its list (which is the only metric requiring a reference forecast),
thereby causing the tests to not result in a ValueError being raised
(which is the point of these tests).