metrics: calculator: handle ending interval label (#297)
* metrics: calculator: improve hourly category labels

If the obs/fx have a `interval_label` of "ending", then the hourly
category labels (for the metrics) should be 1-24 instead of 0-23.

* Move the `closed = ...` outside the loops

Since we only need to call the `closed = ...` look-up once, the line
should be outside the loops to prevent wasted computation.

* calculator: move "hour" category if into if-else

* metrics: calculator: force groupby to match interval_label

Force `groupby` to be consistent with the expected behavior of
`interval_label`. Specifically, if `interval_label == ending`, then the
last interval should be included in the bin (i.e. `(2019-01-01 00:00:00,
2019-01-01 01:00:00]`, not `[2019-01-01 00:00:00, 2019-01-01 01:00:00)`).

* docs/whatsnew: handle ending interval label in metrics

Note that the `metrics.calculator` `groupby` category code now respects
`interval_label == ending` (instead of only `interval_label ==
beginning`).

* reports/figures: offset xticks if hour category

For the metric plots, offset the x-axis ticks for the "hour of the day"
figures such that the bars are between the x-axis ticks (instead of the
default of center aligned over the ticks).

* metrics: calculator: deepcopy fx and obs

If `interval_label == ending`, use a deepcopy of `obs` and `fx` to apply
the 1 nanosecond offset (for correcting the interval selection in
`groupby`).

* Apply groupby hack to `df`, not `obs` and `fx`

* calculator: remove unnecessary `copy` import

* Check that `interval_label` of `ref_fx` matches `obs`/`fx`

* Add test for `interval_label` mismatch in calculator

* Correct flake8 error

* test `interval_label == ending`

* More flake8 corrections

* calculator: simply interval_label binning check

* tests: correct indentation/whitespace

* tests/calculator: run raises example after regular

* tests/calculator: remove unnecessary `return`

* tests/calculator: try to simplify interval_label test

* tests/calculator: further simplify

* tests/calculator: test both interval_label values

* tests/calculator: use `pytest.mark.xfail`

Use `pytest.mark.xfail` for `interval_label` scenarios which should
fail (rather than explicitly checking that the function raises the
correct error/exception type).

* Remove unused `import datetime`

* Specify what error is raised in the failed tests

* Call `calculate_deterministic_metrics()` directly

To increase codecov, try calling `calculate_deterministic_metrics()`
directly instead of `calculate_metrics()`.

* Revert to using `calculate_metrics()`

Revert to calling `calculate_metrics()` since changing to
`calculate_deterministic_metrics()` didn't fix the codecov issue.

* Forgot to update kwarg name change between functions

* tests/calculator: use `strict=True` in pytest.mark.xfail

* tests/calculator: simplify cases

* Switch back to explicit fx/obs generation

To help with debugging, switch back to explicitly defining the fx and
obs objects. Also, it seems there's some sort of bug where the global
variable `DETERMINISTIC_METRICS` doesn't include forecast skill (s) in
its list (which is the only metric requiring a reference forecast),
thereby causing the tests to not result in a ValueError being raised
(which is the point of these tests).

* Switch back to simpler fx/obs generation

* Switch back to `calculator.calculate_metrics()`

* Remove datetime import

* figures: use `dodge()` to tweak bar alignment

Use bokeh's `dodge()` function to tweak the bar alignment in `vbar()`
such that (if plotting hourly results) the bars are aligned between the
x-axis ticks.

* Add tests for interval_label behavior

Add tests to confirm that the `calculate_deterministic_metrics()`
function gives the correct behavior depending on the `interval_label`
(ending vs beginning).

* Correct flake8 line length error