metrics: clarify `interval_label` mismatch behavior
Add a note to the docstring of `resample_and_align()` regarding the
behavior in the case of the `interval_label` not being the same between
the forecast and observation. Also revise tests to explicitly check this
behavior.

Note: this behavior is likely to be revised in the future; the goal of
this commit is to document the current behavior as a first step towards
revising the behavior.