metrics/calculator/tests: test timezone changes
Revise the tests to also check that the metrics per category respect the
timezone of the report (e.g. if the fx/obs are originally in Central
time, but the report is in Pacific, then the RMSE per hour of the day
should be given with the hours converted to Pacific time).

Note: the timezone is provided via a `ReportMetaData` object, which is
used when calling the `metrics` code.