metrics: test timezone behavior (#314)
* metrics/preprocessing: test resample with timezone diff

Test `resample_and_align()` to ensure that the local timezone conversion
is correct when there is a difference between the local timezone (which
will be taken from the `ReportMetaData` object) and the timezone of the
fx/obs pair.

* metrics/calculator/tests: test timezone changes

Revise the tests to also check that the metrics per category respect the
timezone of the report (e.g. if the fx/obs are originally in Central
time, but the report is in Pacific, then the RMSE per hour of the day
should be given with the hours converted to Pacific time).

Note: the timezone is provided via a `ReportMetaData` object, which is
used when calling the `metrics` code.