metrics/event: handle divide-by-zero and use np.logical_and()
Handle divide-by-zero errors by adding a if-else statement to the return
of the functions. Also switch to use the numpy `logical_and()` function
for determining the number of True Positives, False Positives, etc. in
order to fix a flake8 error (which complained that checking for
True/False elements should be done with "if cond is True", not "if cond
== True"). At the moment, there doesn't seem to be a downside or issue
with using `np.logical_and()`, but this may need revision in the future
if we discover a bug/edge case that invalidates the approach.