metrics/event: add tests for full coverage
Add tests for the case where the denominator of the event metrics is
zero for all metrics. Before, only some of the metrics had explicit
tests for this case (which normally would cause a divide-by-zero error,
but does not anymore thanks to the added if-else statement in the
function return statements).