metrics: add event forecast metrics (#348)
* metrics: outline event metric functions and tests

* metrics/event: first draft of functions

* metrics/event: handle divide-by-zero and use np.logical_and()

Handle divide-by-zero errors by adding a if-else statement to the return
of the functions. Also switch to use the numpy `logical_and()` function
for determining the number of True Positives, False Positives, etc. in
order to fix a flake8 error (which complained that checking for
True/False elements should be done with "if cond is True", not "if cond
== True"). At the moment, there doesn't seem to be a downside or issue
with using `np.logical_and()`, but this may need revision in the future
if we discover a bug/edge case that invalidates the approach.

* metrics/event: add working tests

* metrics/event: raise error in _event2count() for no data

Raise a `RuntimeError` in the `_event2count()` function for cases where
there is no observation or forecast timeseries data. May also need to
check that the obs and fx have the same length, but that might be
something better placed inside the `metrics.preprocess` module.

* metrics/event: add tests for full coverage

Add tests for the case where the denominator of the event metrics is
zero for all metrics. Before, only some of the metrics had explicit
tests for this case (which normally would cause a divide-by-zero error,
but does not anymore thanks to the added if-else statement in the
function return statements).

* docs/api: add metrics.event functions

* docs/whatsnew/1.0.0b5: add event metrics to API changes

* metrics/event: add `_MAP`, `__all__`, etc.

* metrics/event: use `np.count_nonzero()` instead of `np.sum()`

Use `np.count_nonzero()` instead of `np.sum()` for computing the number
of True Positives, False Positives, etc.

* Misc fixes

Correct a typo in the docstring for CSI and simplify code for ensuring
inputs are array-like.

* metrics/event: check len(fx) == len(obs)

* metrics/event: add comment on mismatch length RuntimeError

Add a note in the `_event2count()` docstring about the RuntimeError also
being raised if the length of the observations and forecasts are not the
same.

* metrics/event: add Raises section to all docstring

Add the Raises section to the docstring for all functions that rely on
`_event2count()`.

* metrics/event: missing Raises from some func docstrings

* docs/whatsnew: use `:py:func:`, not `:py:mod:`