template: replace `scatter()` with `event_histogram()`
Revise the template so that if the report is for an event forecast,
showing an event histogram (i.e. counts for each outcome: true positive,
true negative, etc.) for each event forecast. The switch occurs inside
the template and fits into the current code flow, thereby minimizing the
number of revisions needed to get the reports rendering properly.