persistence: add outline of ref load forecast
Add outline of the `persistence_hourofday_dayofweek()` function, which
will serve as a baseline/reference load forecast. The basic idea is
similar to a Smart Persistence, but instead of a clear-sky model, the
method uses a dynamically updated trend of the load that is computed as
a function of hour of day (HOD) and day of week (DOW). For variables
such as load, correcting for the HOD and DOW get you pretty far in terms
of baseline performance, while requiring minimal assumptions.