Add `persistence_dayofweek` as reference load forecast method
Add a day of week persistence forecast method that uses data from the
same day of the week from a prior week (e.g. use the previous Monday to
predict the current Monday). The function is effectively just a wrapper
around `persistence_interval`, but is still tested as a separate
function.