Switch from `load` to `net_load`
Since the scope of Solar Forecasting 2 only includes net load forecast
evaluations (not net load and load), switching to supporting `net_load`
as the variable for weekahead persistence forecasts (instead of `load`).
We may come back to this later (i.e. to make it so the Arbiter supports
both `net_load` and `load`), but it's best to start with the one that is
explicitly mentioned by the scope of work.

Related: there may be future questions about whether net load should be
in units of power [MW] or energy [MWh]. It's probably best to just keep
it as power [MW] since 1) that's what other variables (e.g. `ac_power`
use) and 2) we (very) likely don't want to go down the rabit hole of
trying to support things like converting units for reports. Also, if we
just keep it as MW only, that doesn't exclude users from uploading MWh
data and then just treating it as MW data (from the perspective of
evaluating forecast performance).