Add reference load forecast method (#392)
* persistence: add outline of ref load forecast

Add outline of the `persistence_hourofday_dayofweek()` function, which
will serve as a baseline/reference load forecast. The basic idea is
similar to a Smart Persistence, but instead of a clear-sky model, the
method uses a dynamically updated trend of the load that is computed as
a function of hour of day (HOD) and day of week (DOW). For variables
such as load, correcting for the HOD and DOW get you pretty far in terms
of baseline performance, while requiring minimal assumptions.

* temporarily add example script

Temporarily add an example script showing the reference (net) load
forecast method. This script will be removed prior to the final PR
approval, but I've added it to (hopefully) make some of the discussion
around implementation details a little bit easier.

* Add sample CAISO net load CSV

* docs/whatsnew: add note on day of week persistence

* docs/api: add `persistence_dayofweek`

* Add `persistence_dayofweek` as reference load forecast method

Add a day of week persistence forecast method that uses data from the
same day of the week from a prior week (e.g. use the previous Monday to
predict the current Monday). The function is effectively just a wrapper
around `persistence_interval`, but is still tested as a separate
function.

* Delete example script

* Delete sample CAISO data (no longer needed)

* remove print statement (leftover from debugging)

* datamodel: add `load` as allowed variable

Starting with just `load`, but may also add `net_load` and/or `demand`
(as an alias for `load`). Also, not 100% sure on units since load can be
reported with units of power [MW] or energy [MWh]. So starting with [MW]
and then we can revise if needed.

* reference_forecasts/utils: add dayofweek data range func

Add the ability to select data start and end matched by day of the week,
which is necessary for the reference day of week persistence forecast
used for load forecasting.

* Adapt `run_persistence()` to support load forecasts

Adapt `run_persistence()` to support the baseline load forecast method
(day of week persistence).

* persistence: remove wrapper func for dayofweek persistence

Remove wrapper function for day of week persistence since it's not 100%
necessary. Instead we just need `run_persistence()` to support
generating day of week persistence for load forecasts.

* utils: switch to week-ahead function

* docs/api: remove dayofweek persistence func

* utils: update weekahead data range select

* run_persistence: use weekahead persistence for non-intraday load forecasts

* whatsnew/1.0.0b5: update for weekahead persistence

* obs wasn't using `variable` from pytest param

* run_persistence: "tends to exhibit" becomes "exhibit"

* Switch from `load` to `net_load`

Since the scope of Solar Forecasting 2 only includes net load forecast
evaluations (not net load and load), switching to supporting `net_load`
as the variable for weekahead persistence forecasts (instead of `load`).
We may come back to this later (i.e. to make it so the Arbiter supports
both `net_load` and `load`), but it's best to start with the one that is
explicitly mentioned by the scope of work.

Related: there may be future questions about whether net load should be
in units of power [MW] or energy [MWh]. It's probably best to just keep
it as power [MW] since 1) that's what other variables (e.g. `ac_power`
use) and 2) we (very) likely don't want to go down the rabit hole of
trying to support things like converting units for reports. Also, if we
just keep it as MW only, that doesn't exclude users from uploading MWh
data and then just treating it as MW data (from the perspective of
evaluating forecast performance).

* main.py: flake8 in docstring

* whatsnew/1.0.0b5: remove from beta doc

* whatsnew: add changes to 1.0.0rc1

* 1.0.0b5: remove newline to match beta5 version