persistence_probabilistic_timeofday: resample obs
For the `persistence_probabilistic_timeofday()`, resample observation
data to match the forecasts before computing the CDF or percentiles for
each time of day. This step is not necessary in the
`persistence_probabilistic()` function as that function simply uses all
observation data and persists the values forward. In contrast,
`persistence_probabilistic_timeofday()` matches by time of day, which
therefore necessitates that the observation data and forecasts have the
same sampling. For example, if the forecast is at a 1-hour interval
length and the observation is 30-minute interval length, then the
observation should be first resampled to 1-hour and then the CDF or
percentiles should be computed. This resampling should match user
expectations for what to do when you have a mistmatch between the
observation and forecast sampling.