codecov fixes: timezone and axis for timeofday
Modify tests to ensure code coverage for 1) the forecast timestamps not
having timezone info (in which case, they get set as the observation
timezone) and 2) incorrect axis passed to the probabilistic persistence
with timeofday matching. Note: there was a test for the incorrect axis
before, but it was getting hit because it hit a failure before getting
to the raise ValueError code at the end of the function.