Add reference probabilistic forecast (#434)
* Outline probabilistic persistence function

Outline a probabilistic persistence forecast function, which will
(likely) be based on the persistence ensemble (PeEn) method.

* Add `persistence_probabilistic()` as ref prob fx

Add `persistence_probabilistic()` as a reference probabilistic forecast,
which is typically called Persistence Ensemble (PeEn) in the literature.

* add working tests

Still a work-in-progress, but added some (preliminary) tests for the
`persistence_probabilistic()` function. Next step is extended the tests
to additional (common) scenarios.

* tests: start pulling out params to run multiple scenarios

* persistence_probabilistic: works with 'x' and 'y' `axis`

Modify the function to handle providing either `constant_values` as
constant variable values ('x') or constant percentiles ('y'), which
matches the datamodel for probabilistic forecasts. Also add some tests
to confirm the behavior. (Although could probably use a few more tests
to confirm the behavior and that the function is robust to common
situations).

Next step is supporting computing the persistence forecast by time of
day, which will (probably) be done via a separate function to keep
things clear. So you call one function is you want to use all
observation data between `data_start` and `data_end`, and then call a
separate function if you want to, e.g., compute the CDF by time of day
and use that corresponding time of day CDF to forecast (e.g. to forecast
9am, use only observations from 9am from the previous N days).

* test for invalid `axis` value

* docs/api: add persistence_probabilistic()

* whatsnew/1.0.0rc1: add note on probabilistic persistence

* Add `persistence_probabilistic_timeofday()`

Add a function for computing a probabilistic persistence forecast that
takes into consideration time of day (e.g. if forecasting for 9am,
compute the empirical CDF based solely on observations at 9am between
`data_start` and `data_end`). Matching by time of day is beneficial for
day-ahead horizons.

* api: list probabilistic persistence with timeofday matching

* whatsnew/1.0.0rc1: list prob pers timeofday

* flake8: missing whitespace after comma in test

* replace `assert` with `raises`

* persistence_probabilistic_timeofday: resample obs

For the `persistence_probabilistic_timeofday()`, resample observation
data to match the forecasts before computing the CDF or percentiles for
each time of day. This step is not necessary in the
`persistence_probabilistic()` function as that function simply uses all
observation data and persists the values forward. In contrast,
`persistence_probabilistic_timeofday()` matches by time of day, which
therefore necessitates that the observation data and forecasts have the
same sampling. For example, if the forecast is at a 1-hour interval
length and the observation is 30-minute interval length, then the
observation should be first resampled to 1-hour and then the CDF or
percentiles should be computed. This resampling should match user
expectations for what to do when you have a mistmatch between the
observation and forecast sampling.

* Resample in `persistence_probabilistic()`

* Add equations to docstrings explain prob persistence

Add equations to the docstrings to help explain the process ("under the
hood") of the probabilistic persistence forecasts for both constant
variable values and constant probabilities.

* docstrings: update reference linking

* docstring: add Raises section

* replace empty ndarray with `pd.Series` of nan values

Not likely to make a noticable difference in timing, but it's at least a
bit cleaner using a nan-valued pd.Series.

* docstring: cite Yang (2019) and Doubleday et al. (2020)

* timeofday: require at least 20 days of data

Switch from 7 days to 20 days as the minimum amount of data required for
the probabilistic persistence with time of day matching, based on the
three papers cited in the function docstrings.

* typo: s/whicht he/which the/g

* check data length with valid data

* check that fx matches obs timezone

* codecov fixes: timezone and axis for timeofday

Modify tests to ensure code coverage for 1) the forecast timestamps not
having timezone info (in which case, they get set as the observation
timezone) and 2) incorrect axis passed to the probabilistic persistence
with timeofday matching. Note: there was a test for the incorrect axis
before, but it was getting hit because it hit a failure before getting
to the raise ValueError code at the end of the function.

* remove tz conversion for fx index

* validate obs/fx timezones match for timeofday prob persistence

Validate that the obs and fx timezones are compatible, including
converting timezones as needed for the timeofday match.