sfa_reference_sites: simplify network_api_id
Simplify the `network_api_id` for the EIA sites as just the
site-specific part of the `series_id` required by the API. For example,
"EBA.CISO-ALL.D.H" is the `series_id` for CAISO's demand observations in
UTC, where "CISO-ALL" is location (all of CAISO) and "D.H" is the
variable + frequency/timezone ("D" = demand, "H" = hourly UTC). So it's
better to just include the "CISO-ALL" part in `sfa_reference_sites.csv`
and then form the `series_id` in the Python code when fetching the data
for each variable.