test init fx: fix with additional `mocker.patch()` lines
The fix to the failing assert was due to the need to have a net load
observation for the test (instead of the `single_observation`). Once I
added a net load `Observation` object, then it was just a matter of
patching a few more `APISession` list functions (for observations,
forecasts and probabilistic forecasts). With that, the test works
properly.

Note: there's probably a more elegant way to define this test and its
related `mocker.patch()` lines, but this does work. So I think it's good
enough for now.