default_forecasts: revise net load template fx
Since net load forecasts for horizons of 1D or longer already switch to
a day of week (7-day ahead) persistence, we just need to have the
day-ahead template fx. The only change is to rename it as "Day of Week
Persistence" since that is 1) more accurate and 2) likely clearer to
anyone who browses these forecasts on the dashboard.