io/fetch: add support for EIA Open Data API (#485)
* io/fetch: add eia.py for EIA Open Data API

* io/fetch/eia: add test

* test for "bad" data

The EIA API states that it returns "null", "w" or "*" for invalid data.

* add reference to docstring for EIA API

* fetch/eia: return as Series not DataFrame

Since the EIA API returns a single time-series/variable for each call,
it makes more sense to return the data as pandas.Series instead of
pandas.DataFrame.

* api: add eia fetch and reference_observations

* docstring: add example of using get_eia_data()

* sfa_reference_sites: add ISOs/RTOs from EIA

* sfa_reference_sites: use `America/<city>` timezones

* io/reference_observations: add eia.py outline

* io/fetch/eia: return DataFrame, not Series

Return a DataFrame (not a Series) in order to match the other sites
(PVDAQ, RTC, etc.).

* sfa_reference_sites: simplify network_api_id

Simplify the `network_api_id` for the EIA sites as just the
site-specific part of the `series_id` required by the API. For example,
"EBA.CISO-ALL.D.H" is the `series_id` for CAISO's demand observations in
UTC, where "CISO-ALL" is location (all of CAISO) and "D.H" is the
variable + frequency/timezone ("D" = demand, "H" = hourly UTC). So it's
better to just include the "CISO-ALL" part in `sfa_reference_sites.csv`
and then form the `series_id` in the Python code when fetching the data
for each variable.

* io/reference_observations/eia: add main functions

* docs/whatsnew/1.0.0rc1: note support for net load from EIA

* io/eia: note variable name for solar generation (NG.SUN.H)

* io/eia: fetch with UTC timestamps

* Use "eia_site_id" in note on available variables

* Add net load persistence to default_forecasts

* default_forecasts: add Week-ahead Persistence

For net load, include a week-ahead, day-ahead and hour-ahead
persistence in the `TEMPLATE_NETLOAD_PERSISTENCE_FORECASTS`. The
week-ahead persistence should perform better than the day-ahead due to
the stronger correlation between day of week (e.g. Mondays look more
like each other than a Monday and a Sunday), but still including both
week-ahead and day-ahead for completeness.

* reference_data: add EIA

* io/fetch/eia: use `pd.to_numeric()` to replace invalid with NaN

* io/fetch/eia: remove unused numpy import

* Use simpler "EIA" name (instead of "EIA Open Data")

Although the API is listed on a page as "Open Data", that's more of a
description of the data than it is an official name of the EIA "database".

* io/reference_observations/eia: typo in forecast template

* io/reference_observations/eia: add more tests

* io/reference_observations/eia: rename test

* test empty df from fetch if no extra params

* io/reference_observations/eia: fetch return UTC

Remove the `tz_localize()` since that should already be in UTC and we
don't want to return in the local timezone.

* tests_ref_eia: test fetch with data and empty dataframes

* docs/reference-observations: add EIA

* sfa_reference_sites: typo in ERCOT name

* sfa_reference_sites: all EIA data is 1-hour resolution

Update the interval_length to 60 (60mins) for the EIA sites as they are
all hourly resolution.

* default_forecasts: keep all net load as interval label ending

* create_nwp_forecasts: return empty if no vars

The `create_nwp_forecasts()` should return empty if there are no
variables.

* test init fx: try patching API

Not working exactly since the assert fails, but pushing now to help with
debugging via GitHub.

* test init fx: fix with additional `mocker.patch()` lines

The fix to the failing assert was due to the need to have a net load
observation for the test (instead of the `single_observation`). Once I
added a net load `Observation` object, then it was just a matter of
patching a few more `APISession` list functions (for observations,
forecasts and probabilistic forecasts). With that, the test works
properly.

Note: there's probably a more elegant way to define this test and its
related `mocker.patch()` lines, but this does work. So I think it's good
enough for now.

* test init fx: clean up parameters onto one line

* default_forecasts: revise net load template fx

Since net load forecasts for horizons of 1D or longer already switch to
a day of week (7-day ahead) persistence, we just need to have the
day-ahead template fx. The only change is to rename it as "Day of Week
Persistence" since that is 1) more accurate and 2) likely clearer to
anyone who browses these forecasts on the dashboard.