quantile_score: revise to use indicator func form
Revise the quantile_score() function use the indicator function form,
which, while not as intuitive as explicitly listing out the if-else
calculation, is more robust to handling either a single `obs, fx,
fx_prob` input or an array of `obs, fx, fx_prob` inputs. But to
counteract the non-intuitive nature of the indicator function form, this
commit also adds notes in the function docstring to explain how the
function behaves for when `obs > fx` and `obs < fx`.