metrics: add CRPSS (#605)
* metrics: add CRPSS

Add skill score version of the CRPS metric, which shows the relative
improvement (in terms of CRPS) between a forecast versus a reference
forecast.

* metrics: add test for negative CRPSS value

Add test for situation where the CRPS of the forecast is worse than the
reference forecast and therefore the CRPSS is negative.

* docs/api: list CRPSS metric (for prob fx)

* docs/whatsnew/1.0.04rc4: add CRPS skill score mention

* tests: flake8 corrections

* tests/calculator: handle CRPSS requiring ref dist fx

Revisions to handle CRPSS requiring reference forecasts that are full
distribution forecasts (not just single probability forecasts). Still a
work-in-progress from these changes are moving things forward.

* CRPS: add checks to ensure forecasts have proper dimensions

Add checks to ensure that the forecasts given to the CRPS function have
the correct dimensions. If there is an issue, raise a ValueError.

* CRPS: more revisions to improve robustness

Revisions to further improve robustness of inputs to CRPS and CRPSS.
Particularly for the reference forecasts.

* calculator tests: revise for CRPSS changes

* CRPSS: modify to avoid divide by zero

Modify the CRPSS calculation to avoid divide by zero (same pattern as in
`deterministic.forecast_skill`).