metrics: test unequal dist for CRPSS (#608)
Add tests for the situation in which the forecast (fx) and reference
forecasts (ref) have unequal distribution, either a) the same number of
CDF intervals but not the same intervals (e.g. [25, 75] and [50, 100])
or b) a different number of CDF intervals (e.g. [25, 50, 75] and [0, 20,
40, 60, 80, 100]). In these two cases, the expected behavior (currently)
is that the CRPS is calculated for both the fx and ref, and then the
CRPSS is calculated without any adjustments.